+++ android/app/src/main/java/com/mulkkam/di/NetworkInjection.kt
@@ -1,7 +1,7 @@
-package com.seulseul.di
+package com.mulkkam.di
 
+import com.mulkkam.BuildConfig
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import com.seulseul.BuildConfig
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
